Change signature of main to something more standard
This changes the signature of the main() function in the challenges and
POVs to be "int main(int cgc_argc, char *cgc_argv[])". The argument
names were chosen to avoid conflicts with challenges that use the names
argc/argv for their own purposes.

This should help out some analysis tools (specifically KLEE) that expect
main() to have at least two arguments.

This commit doesn't touch challenges that already had arguments to main.
Such challenges need a bit more careful handling, since they expect the
address of the flag page to be passed in the first argument.